adding fixing of permissions to be separate multiprocessing tasks, and fixing progress bars!
