+++ libexec/python/docker/Makefile.am
@@ -1,6 +1,6 @@
 scriptlibexecdir = $(libexecdir)/singularity/python/docker
 
-dist_scriptlibexec_DATA = api.py main.py __init__.py
+dist_scriptlibexec_DATA = api.py main.py tasks.py __init__.py
 
 MAINTAINERCLEANFILES = Makefile.in
