+++ libexec/python/shub/api.py
@@ -128,7 +128,8 @@ class SingularityApiConnection(ApiConnection):
 
         # Download image file atomically, streaming
         image_file = self.download_atomically(url=url,
-                                              file_name=image_file)
+                                              file_name=image_file,
+                                              show_progress=True)
 
         if extract == True:
             if not bot.is_quiet():
